requirements.txt
# Core dependencies
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0
python-dotenv==1.0.0

# Data processing and storage
numpy==1.24.3
pandas==2.1.0
faiss-cpu==1.7.4
chromadb==0.4.7
weaviate-client==3.24.1

# Machine Learning and NLP
scikit-learn==1.3.0
transformers==4.33.2
torch==2.0.1
sentence-transformers==2.2.2

# Graph processing
networkx==3.1
matplotlib==3.7.2
seaborn==0.12.2

# Document processing
beautifulsoup4==4.12.2
scholarly==1.7.11
arxiv==1.4.8

# LangChain and LLM
langchain==0.0.292
openai==0.28.0

# Experiment running
jupyter==1.0.0
nbformat==5.9.2
nbconvert==7.8.0
docker==6.1.3

# Async support
aiohttp==3.8.5
asyncio==3.4.3

# Logging and monitoring
python-json-logger==2.0.7
